+++ src/components/plip/metrics-info.js
@@ -11,6 +11,7 @@ import {
 } from "ramda";
 
 import {
+  countTimingFunction,
   formatNumber,
   remainingDays,
   signatureEnabled,
@@ -21,8 +22,6 @@ import locale from "../../locales/pt-BR";
 import { MKProgress } from "react-native-material-kit";
 import AnimateNumber from "react-native-animate-number";
 
-// slow start, slow end
-const timingFunction = (interval, progress) => interval * (1 - Math.sin(Math.PI*progress)) * 10;
 
 const plipProgress = ({ signaturesRequired, signaturesCount }) => {
   if (!signaturesRequired || signaturesRequired == 0 || !signaturesCount) {
@@ -62,15 +61,39 @@ const renderPlipFinished = () => {
   );
 };
 
-const MetricsInfo = ({ signaturesRequired, signaturesCount, finalDate }) => {
-  const enabled = signatureEnabled({ finalDate });
+const MetricsInfo = ({
+  isRemainingDaysEnabled,
+  signaturesRequired,
+  signaturesCount,
+  totalSignaturesRequired,
+  finalDate,
+}) => {
+  const canSign = signatureEnabled({ finalDate });
+  const showGoal = canSign && !isRemainingDaysEnabled;
+
   return (
     <View>
       <View style={styles.infoContainer}>
-        <TargetPercentage signaturesRequired={signaturesRequired} signaturesCount={signaturesCount} />
-        <SignaturesCount signaturesCount={signaturesCount} />
-        {enabled && <RemainingDays finalDate={finalDate} />}
-        {!enabled && renderPlipFinished()}
+        <View style={styles.infoContainerRow}>
+          <TargetPercentage
+            signaturesRequired={signaturesRequired}
+            signaturesCount={signaturesCount}
+            append={showGoal ? "*" : ""}
+          />
+          <SignaturesCount
+            canSign={canSign}
+            signaturesCount={signaturesCount}
+            signaturesRequired={signaturesRequired}
+            showGoal={showGoal}
+          />
+          {canSign && isRemainingDaysEnabled && <RemainingDays finalDate={finalDate} />}
+          {!canSign && renderPlipFinished()}
+        </View>
+
+        {
+          showGoal &&
+            <Text style={styles.finalGoalText}>* Nossa meta final é de {formatNumber(totalSignaturesRequired)} assinaturas</Text>
+        }
       </View>
 
       <Progress signaturesRequired={signaturesRequired} signaturesCount={signaturesCount} />
@@ -79,9 +102,11 @@ const MetricsInfo = ({ signaturesRequired, signaturesCount, finalDate }) => {
 };
 
 MetricsInfo.propTypes = {
-  finalDate: PropTypes.string,
+  finalDate: PropTypes.string.isRequired,
+  isRemainingDaysEnabled: PropTypes.bool,
   signaturesCount: PropTypes.number.isRequired,
-  signaturesRequired: PropTypes.number,
+  signaturesRequired: PropTypes.number.isRequired,
+  totalSignaturesRequired: PropTypes.number.isRequired,
 };
 
 export default MetricsInfo;
@@ -98,11 +123,11 @@ const Progress = ({ signaturesRequired, signaturesCount }) => {
 };
 
 Progress.propTypes = {
-  signaturesCount: PropTypes.number,
-  signaturesRequired: PropTypes.number,
+  signaturesCount: PropTypes.number.isRequired,
+  signaturesRequired: PropTypes.number.isRequired,
 };
 
-const TargetPercentage = ({ signaturesRequired, signaturesCount }) => {
+const TargetPercentage = ({ append, signaturesRequired, signaturesCount }) => {
   const percentage = progressPercentage({ signaturesRequired, signaturesCount });
 
   return (
@@ -110,38 +135,55 @@ const TargetPercentage = ({ signaturesRequired, signaturesCount }) => {
       <AnimateNumber
         value={percentage}
         countBy={2}
-        timing={timingFunction}
+        timing={countTimingFunction}
         formatter={val => `${val}%`}
         style={styles.infoPercentageText}
       />
-      <Text style={styles.infoPercentageSubtitle}>da meta atual</Text>
+      <Text style={styles.infoPercentageSubtitle}>da meta atual{append}</Text>
     </View>
   );
 };
 
 TargetPercentage.propTypes = {
-  signaturesCount: PropTypes.number,
-  signaturesRequired: PropTypes.number,
+  append: PropTypes.string,
+  signaturesCount: PropTypes.number.isRequired,
+  signaturesRequired: PropTypes.number.isRequired,
 };
 
-const SignaturesCount = ({ signaturesCount }) => {
-  const count = signaturesCount || 0;
+const SignaturesCount = ({ canSign, showGoal, signaturesCount, signaturesRequired }) => {
+  const count = signaturesCount;
+  const goal = signaturesRequired;
+
+  const CountView = () =>
+    <AnimateNumber
+      value={count}
+      timing={countTimingFunction}
+      formatter={formatNumber}
+      style={styles.infoText}
+    />;
 
   return (
-    <View>
-      <AnimateNumber
-        value={count}
-        timing={timingFunction}
-        formatter={formatNumber}
-        style={styles.infoText}
-      />
-      <Text style={styles.infoTextSubtitle}>já assinaram</Text>
+    <View style={[(showGoal ? { marginLeft: 5 } : {})]}>
+      {
+        showGoal &&
+          <View style={{flexDirection: "row", alignItems: "flex-start"}}>
+            <CountView />
+            <Text style={[styles.infoTextSubtitle, { alignSelf: "center", marginLeft: 5 }]}>de</Text>
+            <Text style={[styles.infoText, { marginLeft: 5 }]}>{formatNumber(goal)}</Text>
+          </View>
+      }
+
+      {!showGoal && <CountView />}
+      <Text style={styles.infoTextSubtitle}>{canSign ? "já assinaram" : "assinaram"}</Text>
     </View>
   );
 };
 
 SignaturesCount.propTypes = {
-  signaturesCount: PropTypes.number,
+  canSign: PropTypes.bool,
+  showGoal: PropTypes.bool,
+  signaturesCount: PropTypes.number.isRequired,
+  signaturesRequired: PropTypes.number.isRequired,
 };
 
 const RemainingDays = ({ finalDate }) => {
@@ -154,7 +196,7 @@ const RemainingDays = ({ finalDate }) => {
 };
 
 RemainingDays.propTypes = {
-  finalDate: PropTypes.string,
+  finalDate: PropTypes.string.isRequired,
 };
 
 
@@ -179,17 +221,25 @@ const infoTextSubtitle = {
 };
 
 const styles = StyleSheet.create({
+  finalGoalText: {
+    marginTop: 10,
+    textAlign: "right",
+
+    ...infoTextSubtitle,
+  },
   full: {
     flex: 1,
   },
   infoContainer: {
+    paddingHorizontal: 12,
+    paddingTop: 10,
+    paddingBottom: 15,
+    backgroundColor: "transparent",
+  },
+  infoContainerRow: {
     flexDirection: "row",
     justifyContent: "space-between",
     alignItems: "flex-start",
-    paddingHorizontal: 12,
-    paddingTop: 10,
-    paddingBottom: 20,
-    backgroundColor: "transparent",
   },
   infoPercentageText: {
     ...infoText,