+++ src/selectors/feature.js
@@ -1,0 +1,7 @@
+import {
+  curry,
+} from "ramda";
+
+export const isFeatureEnabled = (feature, state) => state.feature[feature];
+
+export const isRemainingDaysEnabled = curry(isFeatureEnabled)("plipRemainingDaysEnabled");